Opt out the reader part
